UPD - Update dependencies and versions in `pyproject.toml` and `poetry.lock`
This commit updates the `pyproject.toml` and `poetry.lock` files to reflect new package versions and dependencies. Notably, it upgrades `sphinx-rtd-theme` to version `3.0.2`, `docutils` to version `0.21.2`, and introduces new packages such as `markdown-it-py` and `mdit-py-plugins`. Additionally, it modifies the Python version requirements for several packages to ensure compatibility with newer versions. These changes aim to enhance the project's documentation capabilities and maintain up-to-date dependencies for improved functionality.